+++ server/src/completion.ts
@@ -18,6 +18,9 @@ export async function getCompletionItems(
 
 	if (isHTMLFile) {
 		let currentHTMLService = await htmlServiceMap.forceGetServiceByDocument(document)
+		if (!currentHTMLService) {
+			return null
+		}
 
 		let fromPart = currentHTMLService.findDetailedPartAt(offset)
 		if (!fromPart) {
@@ -28,6 +31,9 @@ export async function getCompletionItems(
 	}
 	else if (isCSSFile) {
 		let currentCSSService = await cssServiceMap.forceGetServiceByDocument(document)
+		if (!currentCSSService) {
+			return null
+		}
 
 		let fromPart = currentCSSService.findDetailedPartAt(offset)
 		if (!fromPart) {